// ddktm-electronics-frontend/frontend/src/app/shared/models/user/user.ts
export class User {
    id: number;
    address: string;
    city: string;
    phone: string;
    name: string;
    introduction: string;
    email: string;
    password: string;
    avatar: string;
    status_id: number;
    sns_id: string;
    sns_type: string;
    sns_data: string;
    deleted_at: any;
    activation_token: string;
    remember_token: string;
    token?: string;
    created_at: any;
    updated_at: any;
}